+++ LMeter/Plugin.cs
@@ -18,7 +18,7 @@ namespace LMeter
     {
         public const string ConfigFileName = "LMeter.json";
 
-        public static string Version { get; private set; } = "0.3.0.1";
+        public static string Version { get; private set; } = "0.3.1.0";
 
         public static string ConfigFileDir { get; private set; } = "";
 
@@ -92,7 +92,14 @@ namespace LMeter
             // Load config
             FontsManager.CopyPluginFontsToUserPath();
             LMeterConfig config = ConfigHelpers.LoadConfig(Plugin.ConfigFilePath);
+
+            // Check if any old configs need to be converted
+            ConfigHelpers.ConvertOldConfigs(config);
+
+            // Refresh fonts
             config.FontConfig.RefreshFontList();
+
+            // Register config
             Singletons.Register(config);
 
             // Initialize Fonts